refactor: 리팩토링 세분화
